+++ src/Exceptions/ContainerException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Mikro\Exceptions;
+
+class ContainerException extends \Exception
+{
+    //
+}