+++ src/Exceptions/EventException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Mikro\Exceptions;
+
+class EventException extends \Exception
+{
+    //
+}