+++ src/Exceptions/JwtException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Mikro\Exceptions;
+
+class JwtException extends \Exception
+{
+    //
+}