+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to `mikro` will be documented in this file
 
-## 1.0.0 - 2022-10-01
+## 1.1.0 - 2022-03-01
+- Exceptions added
+- `dont_report` method added to Error component
+- Fix on CurlError exception message
+- `isServerError`, `isClientError`, `isFailed`, `isOk`, `isRedirect` and `errorOnFail` methods added to Curl component
+- Some PHP 8 improvements
 
+## 1.0.0 - 2022-10-01
 - initial release
 - It is not compatible with the previous version.